Implement user restrictions

+++ app/Http/Controllers/Tenants/UsersController.php
@@ -2,8 +2,10 @@
 
 namespace App\Http\Controllers\Tenants;
 
+use App\Enums\UserType;
 use App\Models\Invitation;
 use App\Models\User;
+use Illuminate\Auth\Access\AuthorizationException;
 use Illuminate\Contracts\View\View;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
@@ -12,43 +14,58 @@ use Illuminate\Http\Response;
 use Illuminate\Support\Facades\DB as FacadesDB;
 use Illuminate\Support\Facades\Hash;
 use Illuminate\Validation\Rule;
+use Illuminate\Validation\Rules\Enum;
 use Illuminate\Validation\ValidationException;
 use Spatie\Permission\Models\Role;
 
 class UsersController extends Controller
 {
-    public function index(): View
+  /**
+   * @throws AuthorizationException
+   */
+  public function index(): View
     {
+      $this->authorize('view user');
         return view('tenants.users.index', [
           'users' => User::with('roles')->paginate(10)
         ]);
     }
 
-    /**
-     * Show the form for creating a new resource.
-     *
-     * @return View
-     */
+  /**
+   * Show the form for creating a new resource.
+   *
+   * @return View
+   */
     public function create(): View
     {
       $roles = Role::pluck('name', 'name')->all();
+      $userTypes = array_filter(array_column(UserType::cases(), 'value'), function ($value) {
+        return $value !== UserType::Tenant->value;
+      });
 
-      return view('tenants.users.create', compact('roles'));
+      return view('tenants.users.create', compact('roles', 'userTypes'));
     }
 
   /**
    * Store a newly created resource in storage.
    *
-   * @param  \Illuminate\Http\Request  $request
+   * @param \Illuminate\Http\Request $request
    * @return \Illuminate\Http\RedirectResponse
+   * @throws AuthorizationException
    */
   public function store(Request $request): \Illuminate\Http\RedirectResponse
   {
+    $this->authorize('create user');
+
     $record =  $request->validate([
       'name' => 'required|string|max:250',
       'email' => 'required|email|max:250|unique:users',
       'username' => 'required|string|max:250',
       'password' => 'required|string|min:8',
+      'type' => [
+        'required',
+        new Enum(UserType::class)
+      ]
     ]);
 
     $record['password'] = Hash::make($request->password);
@@ -64,9 +81,12 @@ class UsersController extends Controller
    *
    * @param User $user
    * @return View
+   * @throws AuthorizationException
    */
   public function show(User $user): View
   {
+    $this->authorize('view user');
+
     return view('tenants.users.show', compact('user'));
   }
 
@@ -90,9 +110,12 @@ class UsersController extends Controller
    * @param Request $request
    * @param User $user
    * @return RedirectResponse
+   * @throws AuthorizationException
    */
   public function update(Request $request, User $user): RedirectResponse
   {
+    $this->authorize('update user', $user);
+
     $request->validate([
       'name' => 'required',
       'email' => [
@@ -116,9 +139,12 @@ class UsersController extends Controller
    *
    * @param User $user
    * @return RedirectResponse
+   * @throws AuthorizationException
    */
   public function destroy(User $user): \Illuminate\Http\RedirectResponse
   {
+    $this->authorize('delete user');
+
     $user->delete();
 
     return redirect()->route('tenants.users.index', tenant('id'))
